+++ scim-ri/src/test/java/com/unboundid/scim/ri/SCIMServerTestCase.java
@@ -1900,7 +1900,7 @@ public abstract class SCIMServerTestCase extends SCIMRITestCase
     final SCIMEndpoint<UserResource> userEndpoint = service.getUserEndpoint();
     int pageSize = 1;
     final Set<String> userIDs = new HashSet<String>();
-    for (long startIndex = 1; startIndex <= NUM_USERS; startIndex += pageSize)
+    for (int startIndex = 1; startIndex <= NUM_USERS; startIndex += pageSize)
     {
       final Resources<UserResource> resources =
           userEndpoint.query("userName sw \"paginationUser\"", null,
@@ -1913,7 +1913,7 @@ public abstract class SCIMServerTestCase extends SCIMRITestCase
     assertEquals(userIDs.size(), NUM_USERS);
 
     // Create some groups.
-    final long NUM_GROUPS = 10;
+    final int NUM_GROUPS = 10;
     for (int i = 0; i < NUM_GROUPS; i++)
     {
       final String cn = "paginationGroup." + i;
@@ -1926,7 +1926,7 @@ public abstract class SCIMServerTestCase extends SCIMRITestCase
            service.getGroupEndpoint();
     pageSize = 3;
     final Set<String> groupIDs = new HashSet<String>();
-    for (long startIndex = 1; startIndex <= NUM_GROUPS; startIndex += pageSize)
+    for (int startIndex = 1; startIndex <= NUM_GROUPS; startIndex += pageSize)
     {
       final Resources<GroupResource> resources =
         groupEndpoint.query("displayName sw \"paginationGroup\"",
@@ -1948,7 +1948,7 @@ public abstract class SCIMServerTestCase extends SCIMRITestCase
     assertEquals(groupIDs.size(), NUM_GROUPS);
 
     // Attempt to fetch resources from a non-existent page.
-    final long startIndex = NUM_GROUPS + 1;
+    final int startIndex = NUM_GROUPS + 1;
     final Resources<GroupResource> resources =
         groupEndpoint.query("displayName sw \"paginationGroup\"",
           new SortParameters("displayName", "ascending"),
